Move to simd.rs
