+++ libafl/Cargo.toml
@@ -40,6 +40,7 @@ default = [
   "serdeany_autoreg",
   "libafl_bolts/xxh3",
   "tui_monitor",
+  "stable_simd",
 ]
 document-features = ["dep:document-features"]
 
@@ -195,6 +196,9 @@ nautilus = [
   "regex-syntax",
   "regex",
 ]
+
+## Use the best SIMD implementation by our [benchmark](https://github.com/wtdcode/libafl_simd_bench)
+stable_simd = ["libafl_bolts/stable_simd"]
 
 [[example]]
 name = "tui_mock"
