+++ libafl/src/feedbacks/mod.rs
@@ -46,6 +46,8 @@ pub mod map;
 pub mod nautilus;
 #[cfg(feature = "std")]
 pub mod new_hash_feedback;
+#[cfg(feature = "stable_simd")]
+pub mod simd;
 #[cfg(feature = "std")]
 pub mod stdio;
 pub mod transferred;
